Split implementations for rms and samples painter
